+++ main.struct.go
@@ -7,41 +7,45 @@ import (
 )
 
 //----- Constants -----
-const version = "1.13.1"
+const version = "1.15.0"
 const appServiceManager = "com.hornbill.servicemanager"
 
 //----- Variables -----
 var (
-	assets            = make(map[string]string)
-	maxLogFileSize    int64
-	SQLImportConf     sqlImportConfStruct
-	Sites             []siteListStruct
-	Groups            []groupListStruct
-	counters          counterTypeStruct
-	configFileName    string
-	configMaxRoutines string
-	configDebug       bool
-	configDryRun      bool
-	configVersion     bool
-	Customers         []customerListStruct
-	TimeNow           string
-	APITimeNow        string
-	startTime         time.Time
-	endTime           time.Duration
-	AssetClass        string
-	AssetTypeID       int
-	BaseSQLQuery      string
-	StrAssetType      string
-	StrSQLAppend      string
-	mutex             = &sync.Mutex{}
-	mutexBar          = &sync.Mutex{}
-	mutexCounters     = &sync.Mutex{}
-	mutexCustomers    = &sync.Mutex{}
-	mutexGroup        = &sync.Mutex{}
-	mutexSite         = &sync.Mutex{}
-	worker            sync.WaitGroup
-	maxGoroutines     = 1
-	logFilePart       = 0
+	assets                 = make(map[string]string)
+	maxLogFileSize         int64
+	SQLImportConf          sqlImportConfStruct
+	Sites                  []siteListStruct
+	Groups                 []groupListStruct
+	counters               counterTypeStruct
+	configFileName         string
+	configMaxRoutines      string
+	configDebug            bool
+	configDryRun           bool
+	configVersion          bool
+	Customers              []customerListStruct
+	TimeNow                string
+	APITimeNow             string
+	startTime              time.Time
+	endTime                time.Duration
+	AssetClass             string
+	AssetTypeID            int
+	BaseSQLQuery           string
+	StrAssetType           string
+	StrSQLAppend           string
+	HInstalledApplications []string
+	blnCMInPolicy          bool
+	blnContractConnect     bool
+	blnSupplierConnect     bool
+	mutex                  = &sync.Mutex{}
+	mutexBar               = &sync.Mutex{}
+	mutexCounters          = &sync.Mutex{}
+	mutexCustomers         = &sync.Mutex{}
+	mutexGroup             = &sync.Mutex{}
+	mutexSite              = &sync.Mutex{}
+	worker                 sync.WaitGroup
+	maxGoroutines          = 1
+	logFilePart            = 0
 )
 
 //----- Structures -----
@@ -69,6 +73,7 @@ type sqlImportConfStruct struct {
 	APIKey                   string
 	InstanceID               string
 	Entity                   string
+	HornbillUserIDColumn     string
 	LogSizeBytes             int64
 	SQLConf                  sqlConfStruct
 	AssetTypes               []assetTypesStruct
@@ -88,9 +93,12 @@ type assetTypesStruct struct {
 }
 
 type assetIdentifierStruct struct {
-	DBColumn     string
-	Entity       string
-	EntityColumn string
+	DBContractColumn string
+	DBSupplierColumn string
+	DBInPolicyColumn string
+	DBColumn         string
+	Entity           string
+	EntityColumn     string
 }
 
 type sqlConfStruct struct {
@@ -112,6 +120,11 @@ type xmlmcResponse struct {
 	State        stateStruct  `xml:"state"`
 }
 
+type xmlmcCreateResponse struct {
+	MethodResult    string      `xml:"status,attr"`
+	PrimaryKeyValue string      `xml:"params>primaryEntityKeyValue"`
+	State           stateStruct `xml:"state"`
+}
 type xmlmcUpdateResponse struct {
 	MethodResult string      `xml:"status,attr"`
 	UpdatedCols  updatedCols `xml:"params>primaryEntityData>record"`
@@ -164,6 +177,93 @@ type xmlmcCustomerListResponse struct {
 	CustomerLastName  string      `xml:"params>lastName"`
 	CustomerID        string      `xml:"params>customerId"`
 	State             stateStruct `xml:"state"`
+}
+type xmlmcCountResponse struct {
+	Params struct {
+		RowData struct {
+			Row []struct {
+				Count string `json:"count"`
+			} `json:"row"`
+		} `json:"rowData"`
+	} `json:"params"`
+	State stateJSONStruct `json:"state"`
+}
+type stateJSONStruct struct {
+	Code      string `json:"code"`
+	Service   string `json:"service"`
+	Operation string `json:"operation"`
+	Error     string `json:"error"`
+}
+
+type xmlmcUserListResponse struct {
+	Params struct {
+		RowData struct {
+			Row []userAccountStruct `json:"row"`
+		} `json:"rowData"`
+	} `json:"params"`
+	State stateJSONStruct `json:"state"`
+}
+type userAccountStruct struct {
+	HUserID     string `json:"h_user_id"`
+	HLoginID    string `json:"h_login_id"`
+	HEmployeeID string `json:"h_employee_id"`
+	HName       string `json:"h_name"`
+	HFirstName  string `json:"h_first_name"`
+	HMiddleName string `json:"h_middle_name"`
+	HLastName   string `json:"h_last_name"`
+	HPhone      string `json:"h_phone"`
+	HEmail      string `json:"h_email"`
+	HMobile     string `json:"h_mobile"`
+	HSnA        string `json:"h_sn_a"`
+	HSnB        string `json:"h_sn_b"`
+	HAttrib1    string `json:"h_attrib_1"`
+	HAttrib2    string `json:"h_attrib_2"`
+	HAttrib3    string `json:"h_attrib_3"`
+	/*	HJobTitle            string `json:"h_job_title"`
+		HLoginCreds          string `json:"h_login_creds"`
+		HClass               string `json:"h_class"`
+		HAvailStatus         string `json:"h_avail_status"`
+		HAvailStatusMsg      string `json:"h_avail_status_msg"`
+		HTimezone            string `json:"h_timezone"`
+		HCountry             string `json:"h_country"`
+		HLanguage            string `json:"h_language"`
+		HDateTimeFormat      string `json:"h_date_time_format"`
+		HDateFormat          string `json:"h_date_format"`
+		HTimeFormat          string `json:"h_time_format"`
+		HCurrencySymbol      string `json:"h_currency_symbol"`
+		HLastLogon           string `json:"h_last_logon"`
+		HSnC                 string `json:"h_sn_c"`
+		HSnD                 string `json:"h_sn_d"`
+		HSnE                 string `json:"h_sn_e"`
+		HSnF                 string `json:"h_sn_f"`
+		HSnG                 string `json:"h_sn_g"`
+		HSnH                 string `json:"h_sn_h"`
+		HIconRef             string `json:"h_icon_ref"`
+		HIconChecksum        string `json:"h_icon_checksum"`
+		HDob                 string `json:"h_dob"`
+		HAccountStatus       string `json:"h_account_status"`
+		HFailedAttempts      string `json:"h_failed_attempts"`
+		HIdxRef              string `json:"h_idx_ref"`
+		HSite                string `json:"h_site"`
+		HManager             string `json:"h_manager"`
+		HSummary             string `json:"h_summary"`
+		HInterests           string `json:"h_interests"`
+		HQualifications      string `json:"h_qualifications"`
+		HPersonalInterests   string `json:"h_personal_interests"`
+		HSkills              string `json:"h_skills"`
+		HGender              string `json:"h_gender"`
+		HNationality         string `json:"h_nationality"`
+		HReligion            string `json:"h_religion"`
+		HHomeTelephoneNumber string `json:"h_home_telephone_number"`
+		HHomeAddress         string `json:"h_home_address"`
+		HBlog                string `json:"h_blog"`
+		HAttrib4             string `json:"h_attrib_4"`
+		HAttrib5             string `json:"h_attrib_5"`
+		HAttrib6             string `json:"h_attrib_6"`
+		HAttrib7             string `json:"h_attrib_7"`
+		HAttrib8             string `json:"h_attrib_8"`
+		HHomeOrg             string `json:"h_home_organization"`
+	*/
 }
 
 //Asset Structs